删除按钮添加提示，景区管理页面调整查询条件

+++ src/views/reOrder/dmq.vue
@@ -321,15 +321,23 @@ export default {
         })
         return
       }
-      deleteDmqTicket(this.rowIdList).then(res => {
-        if (res.data.code !== 200) {
-          return this.$message.error(res.data.error)
-        }
-        _this.getTicketList(this.paramForm)
-        this.$message({
-          message: '操作成功',
-          type: 'success'
+      this.$confirm('确定要删除吗?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteDmqTicket(this.rowIdList).then(res => {
+          if (res.data.code !== 200) {
+            return this.$message.error(res.data.error)
+          }
+          _this.getTicketList(this.paramForm)
+          this.$message({
+            message: '操作成功',
+            type: 'success'
+          })
         })
+      }).catch(() => {
+        return
       })
     },
     // 列表全选按钮
@@ -342,15 +350,23 @@ export default {
     // 每行的删除按钮
     handleDeleteSingle(id) {
       const _this = this
-      deleteDmqTicket([id]).then(res => {
-        if (res.data.code !== 200) {
-          return this.$message.error(res.data.error)
-        }
-        _this.getTicketList(this.paramForm)
-        this.$message({
-          message: '操作成功',
-          type: 'success'
+      this.$confirm('确定要删除吗?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        deleteDmqTicket([id]).then(res => {
+          if (res.data.code !== 200) {
+            return this.$message.error(res.data.error)
+          }
+          _this.getTicketList(this.paramForm)
+          this.$message({
+            message: '操作成功',
+            type: 'success'
+          })
         })
+      }).catch(() => {
+        return
       })
     },
     // 每行的修改按钮
